# server/api/models/Frame.ts
/* tslint:disable */
/* eslint-disable */
/**
 * EPYC API
 * epyc API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import {
    FramePlayData,
    FramePlayDataFromJSON,
    FramePlayDataFromJSONTyped,
    FramePlayDataToJSON,
    Person,
    PersonFromJSON,
    PersonFromJSONTyped,
    PersonToJSON,
} from './';

/**
 * 
 * @export
 * @interface Frame
 */
export interface Frame {
    /**
     * 
     * @type {Person}
     * @memberof Frame
     */
    person: Person;
    /**
     * 
     * @type {FramePlayData}
     * @memberof Frame
     */
    playData: FramePlayData;
}

export function FrameFromJSON(json: any): Frame {
    return FrameFromJSONTyped(json, false);
}

export function FrameFromJSONTyped(json: any, ignoreDiscriminator: boolean): Frame {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'person': PersonFromJSON(json['person']),
        'playData': FramePlayDataFromJSON(json['playData']),
    };
}

export function FrameToJSON(value?: Frame | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'person': PersonToJSON(value.person),
        'playData': FramePlayDataToJSON(value.playData),
    };
}