# server/api/models/FramePlayTitleRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * EPYC API
 * epyc API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface FramePlayTitleRequest
 */
export interface FramePlayTitleRequest {
    /**
     * 
     * @type {string}
     * @memberof FramePlayTitleRequest
     */
    title: string;
}

export function FramePlayTitleRequestFromJSON(json: any): FramePlayTitleRequest {
    return FramePlayTitleRequestFromJSONTyped(json, false);
}

export function FramePlayTitleRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): FramePlayTitleRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'title': json['title'],
    };
}

export function FramePlayTitleRequestToJSON(value?: FramePlayTitleRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'title': value.title,
    };
}